requirements.txt
pypdf>=5.1.0
python-docx>=1.1.2
rich>=13.9.4
